[S08P12A809-174] Update: 대나무 삭제

+++ back/imfine/src/main/java/com/idle/imfine/data/dto/leaf/request/RequestLeafDto.java
@@ -14,11 +14,11 @@ import lombok.ToString;
 @ToString
 public class RequestLeafDto {
     private String writerId;
-    private int bambooId;
+    private Long bambooId;
     private String content;
 
     @Builder
-    public RequestLeafDto(int bambooId, String content) {
+    public RequestLeafDto(long bambooId, String content) {
         this.bambooId = bambooId;
         this.content = content;
     }
